Revert "Flask day"
