+++ README.md
@@ -1,0 +1,3 @@
+# Portfolio
+
+Welcome to my portfolio! Check it out in action at the [deployment here](https://gumdropsteve.github.io/portfolio_website/ds) or run locally with the Flask app [server.py](https://github.com/gumdropsteve/portfolio_website/blob/master/server.py)
