+++ docs/README.md
@@ -1,0 +1,1 @@
+This folder contains the HTML pages being used at [winstonrobson.com](http://winstonrobson.com)
