--- flask_day/README.md
@@ -1,8 +1,0 @@
-# Flask Day
-
-This is the branch of the repo for Winston's instruction of Flask Day at Galvanize.
-
-For the main page of this repo click [here](https://github.com/gumdropsteve/portfolio_website) or the main deployment of this repo click [here](https://gumdropsteve.github.io/portfolio_website/ds). 
-
-### Useful Info
-[Spin up virtual machine on AWS and use it to host your Flask server](http://bit.ly/aws-with-flask-friend)
