--- flask_day/docs/README.md
@@ -1,1 +1,0 @@
-These files are for deploying to GitHub Pages. Like this [example](https://gumdropsteve.github.io/portfolio_website/).
