--- flask_day/docs/home.html
@@ -1,17 +1,0 @@
-<!-- What the page is called in the tab bar  -->
-<!-- We will add out stylsheet here later -->
-<head>
-    <link rel="stylesheet" href="/static/home.css" />
-    <title> Home Page</title>
-</head>
-<!-- Put some text on the page -->
-<body>
-  <h1 id='HomeHeader' >Welcome to my webpage!</h1>
-  <p id='HomeHeader2'>
-    My name is Elliott and this is the beginning of my webpage
-  </p>
-  <span>
-    Here is a span
-  </span>
-</body>
-</html>
