--- flask_day/static/README.md
@@ -1,1 +1,0 @@
-These files are for running Flask.
