--- flask_day/static/home.css
@@ -1,12 +1,0 @@
-#HomeHeader {
-background-color: yellow;
-border: 3px dashed green;
-margin: 25px;
-padding: 25px;
-}
-#HomeHeader2 {
-background-color: yellow;
-border: 3px dashed green;
-margin: 25px;
-padding: 25px;
-}
