--- flask_day/static/mpg.js
@@ -1,55 +1,0 @@
-$(document).ready(function(){
-  console.log('document is ready');
-
-  $('#inference').click(async function(){
-    console.log('button was clicked');
-
-    const cylinders = parseFloat($('#cylinders').val());
-    const horsepower = parseFloat($('#horsepower').val());
-    const weight = parseFloat($('#weight').val());
-    const data = {
-      cylinders,
-      horsepower,
-      weight
-    }
-    console.log(data)
-
-    const response = await $.ajax('/inference',{
-      data: JSON.stringify(data),
-      method: "post",
-      contentType: "application/json"
-    })
-    console.log(response)
-    $('#mpg').val(response.prediction)
-
-  })
-  $('#scatter-button').click(async function(){
-    console.log($('#graph1')[0])
-    console.log('scatter button clicked')
-    const response = await $.ajax('/plot')
-    console.log(response)
-    const mpg = response.map(a => a[0])
-    const weight = response.map(a => a[1])
-    console.log(mpg)
-    const trace = [{
-      x:weight,
-      y:mpg,
-      mode:'markers',
-      type:'scatter'
-    }];
-    const layout = {
-      xaxis:{
-        title:'Weight'
-      },
-      yaxis:{
-        title:'mpg'
-      },
-      title:'Scatter MPG vs Weight',
-      width:700,
-      height:300
-    }
-    Plotly.plot($('#graph1')[0],trace,layout)
-    $('#scatter-button').hide()
-  })
-
-})
