--- flask_day/templates/mpg.html
@@ -1,51 +1,0 @@
-<html>
-
-<head>
-    <link rel="stylesheet" href="/static/mpg.css" />
-    <title>Mpg</title>
-</head>
-<body>
-
-  <img src="http://www.gotmpg.com/images/got_mpg_bumpersticker1.jpg"/>
-  <ul>
-    <li>
-      <a href="http://google.com">Google</a>
-    </li>
-    <li>
-      <a href="/">Home</a>
-    </li>
-  </ul>
-
-  <h1> Welcome to Machine Learning and Data Science v1.0</h1>
-
-  <div id="controls">
-    <div>
-      <label>Cylinders</label>
-      <input id="cylinders" type="number" value="6" />
-    </div>
-    <div>
-      <label>Horsepower</label>
-      <input id="horsepower" type="number" value="150" />
-    </div>
-    <div>
-      <label>Weight</label>
-      <input id="weight" type="number" value="1500" />
-    </div>
-  </div>
-  <div id='inference-button'>
-    <div>
-          <button id="inference">Predict MPG</button>
-    </div>
-    <div>
-          <input id="mpg" type="text" readonly />
-    </div>
-  </div>
-
-<button id="scatter-button">Scatter Plot </button>
-<div id='graph1'><div>
-<script src="https://cdnjs.cloudflare.com/ajax/libs/plotly.js/1.35.2/plotly.min.js"></script>
-<script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
-<script src="/static/mpg.js"></script>
-</body>
-
-</html>
