+++ static/about.css
@@ -1,0 +1,13 @@
+#AboutHeader{
+    background-color: yellow;
+    border: 3px dashed green;
+    margin: 25px;
+    padding: 25px;
+}
+#AboutHeader2{
+    background-color: green;
+    border: 3px dashed yellow;
+    margin: 25px;
+    padding: 25px;
+}
+
