+++ static/articles.css
@@ -1,0 +1,13 @@
+#ArticlesHeader{
+    background-color: yellow;
+    border: 3px dashed green;
+    margin: 25px;
+    padding: 25px;
+}
+#ArticlesHeader2{
+    background-color: green;
+    border: 3px dashed yellow;
+    margin: 25px;
+    padding: 25px;
+}
+
