+++ static/home.css
@@ -1,0 +1,20 @@
+#HomeHeader{
+    background-color: navy;
+    color: white
+    border: 3px white;
+    margin: 25px;
+    padding: 25px;
+}
+h1 {
+    font-family: 'Georgia', serif;
+    color: white;
+    background-color: #0D122B;
+    margin: 25px;
+    padding: 25px;
+}
+#HomeHeader2{
+    background-color: green;
+    border: 3px dashed yellow;
+    margin: 25px;
+    padding: 25px;
+}
