+++ static/mpg.css
@@ -1,3 +1,8 @@
+img{
+    height: 150px;
+    width: 150px;
+}
+
 body {
     background-color: #cccccc;
 }
