+++ templates/old/old_home.html
@@ -1,0 +1,24 @@
+<html>
+<head>
+    <title> Winstons Home Page </title>
+    <link rel='stylesheet' href='static/home.css'>
+    
+<head>
+
+<body>
+    <h1>Welcome to Winston's Portfolio</h1>
+    <img src='https://avatars3.githubusercontent.com/u/43570913?s=460&v=4' height="200" width="200"/>
+    <img src="https://media.licdn.com/dms/image/C5603AQGm3jKMOArRAQ/profile-displayphoto-shrink_200_200/0?e=1563408000&v=beta&t=2Ph1GJtf4wGyAtV-VlLGrnxapBhOmsFaIUsMDCZTRQI"/>
+    <strong><nav>
+        <ul class="menu">
+            <li><a href="{{ url_for('home') }}">Home</a></li>
+            <li><a href="{{ url_for('about') }}">About</a></li>
+            <li><a href="{{ url_for('articles') }}">Articles</a></li>
+            <li><a href="{{ url_for('projects') }}">Projects</a></li>
+        </ul>
+    </nav></strong> 
+    <p id='HomeHeader2'> My name is Winston & welcome to my webpage </p>
+
+    <span>Here is a span</span>
+</body>
+</html>
