+++ templates/old/projects.html
@@ -1,0 +1,29 @@
+<html>
+<head>
+    <title> Winstons Home Page </title>
+    <link rel='stylesheet' href='static/projects.css'>
+    
+<head>
+
+<body>
+    <ul>
+        <li>
+            <a href='/'> Home</a>
+        </li>
+        <li>
+            <a href='/about'>About</a>
+        </li>
+        <li>
+            <a href='/articles'>Articles</a>
+        </li>
+        <li>
+            <a href='/projects'>Projects</a>
+        </li>
+    </ul>    
+    <h1 id='ProjectsHeader'> Welcome to my webpage </h1>
+
+    <p id='ProjectsHeader2'> My name is Winston & welcome to my webpage </p>
+
+    <span>Here is a span</span>
+</body>
+</html>
